refactor: update color scheme to match new design for icons and manifest
Update color scheme to align with new icon design

Refactors the color palette across the application to match the new design for icons and the manifest.

Changes include updating the theme color to a purple shade (#6750a4) and adjusting various background elements to enhance the overall visual consistency.

Improvements ensure a cohesive look and feel that aligns with the branding of the TGATB Habit Tracker app.